test: enable ExtensionUtils-integ-test and stability fix
