+++ test/UnitTestSuite.js
@@ -84,8 +84,8 @@ define(function (require, exports, module) {
     require("spec/DocumentManager-integ-test");
     require("spec/DragAndDrop-integ-test");
     require("spec/EditorOptionHandlers-integ-test");
+    require("spec/ExtensionUtils-integ-test");
     // require("spec/ExtensionInstallation-test");
-    // require("spec/ExtensionUtils-test");
     // require("spec/FindInFiles-test");
     // require("spec/InlineEditorProviders-test");
     // require("spec/InstallExtensionDialog-test");
